Change structure and add tests
The structure of the project was greatly changed so tests could be
added. It should be more clear now. 'rcolgen' is now a library with
root at lib.rs, while main.rs is the main binary of the application.

Two tests for the hex conversion method have been made. Their results
are in comments in the code. The errors are at edge cases 0 and
multiples of 16.